requirement.txt
mysql-connector-python
pymysql
sqlalchemy
pandas
openpyxl
tqdm
faker
colorama
